Inline headers from vecgeom
